+++ src/Manager/ProductDataManager.php
@@ -10,6 +10,7 @@ namespace HeimrichHannot\IsotopeBundle\Manager;
 
 use Contao\Controller;
 use HeimrichHannot\IsotopeBundle\Model\ProductDataModel;
+use HeimrichHannot\IsotopeBundle\Model\ProductModel;
 use Isotope\Model\Product;
 
 class ProductDataManager
@@ -18,6 +19,11 @@ class ProductDataManager
      * @var array
      */
     protected $productDataFields;
+
+    /**
+     * @var ProductDataModel[]
+     */
+    protected $productDataModelCache = [];
 
     /**
      * Returns all product data fields.
@@ -35,8 +41,9 @@ class ProductDataManager
             $metaFields = [];
             foreach ($fields as $key => $field) {
                 if (true !== $field['eval']['skipProductPalette']) {
-                    unset($field['sql']);
+//                    unset($field['sql']);
                     $field['save_callback'][] = ['huh.isotope.listener.callback.product', 'saveMetaFields'];
+                    $field['load_callback'][] = ['huh.isotope.listener.callback.product', 'getMetaFieldValue'];
                     $metaFields[$key] = $field;
                 }
             }
@@ -50,17 +57,26 @@ class ProductDataManager
      * Returns the product data for a product.
      * If no product data is available, a new instance will be returned.
      *
-     * @param Product|int $product
+     * @param Product|ProductModel|int $product The product model or id
      *
      * @return ProductDataModel
      */
-    public function getProductDataByProduct($product)
+    public function getProductData($product)
     {
-        $pid = is_int($product) ?: $product->id;
+        $pid = is_numeric($product) ? (int) $product : (int) $product->id;
+
+        if (array_key_exists($pid, $this->productDataModelCache)) {
+            return $this->productDataModelCache[$pid];
+        }
 
         $productData = ProductDataModel::findOneBy('pid', $pid);
         if (!$productData) {
             $productData = new ProductDataModel();
+            $productData->pid = $pid;
+            $productData->tstamp = time();
+            $productData->dateAdded = time();
+        } else {
+            $this->productDataModelCache[$productData->pid] = $productData;
         }
 
         return $productData;